Upgrade specialized mul to malicious
