+++ src/protocol/modulus_conversion/mod.rs
@@ -1,4 +1,3 @@
 mod convert_shares;
-mod specialized_mul;
 
 pub use convert_shares::{convert_one_bit, convert_shares_for_a_bit};
